Se completa el algoritmo de migrar usuarios y calificaciones en las estrategias EstrategiaMigrarUsuarios y EstrategiaImportarCalificaciones

+++ src/main/java/org/moodle/api/client/moodle_rest_client/application/services/gradereport/CourseGradeService.java
@@ -13,7 +13,7 @@ public class CourseGradeService implements GetCourseGradesUseCase {
     private final CourseGradePort courseGradePort;
 
     @Override
-    public CourseGrades getCourseGrades(Integer userId) {
+    public CourseGrades getCourseGrades(Long userId) {
         return courseGradePort.getCourseGrades(userId);
     }
 }
